converted all cli files to ts
