+++ cli/package.json
@@ -1,21 +1,25 @@
 {
-    "name": "@storybook/aem-cli",
-    "version": "0.0.1",
-    "description": "CLI tool for Storybook AEM",
-    "main": "src/index.js",
-    "files": [
-      "src/index.js"
-    ],
-    "scripts": {
-        "prepare": "node ../../scripts/prepare.js"
-    },
-    "keywords": [
-      "storybook",
-      "aem",
-      "cli"
-    ],
-    "dependencies": {
-    },
-    "devDependencies": {
-    }
-  }
+  "name": "@storybook/aem-cli",
+  "version": "0.0.1",
+  "description": "CLI tool for Storybook AEM",
+  "main": "src/index.js",
+  "files": [
+    "src/index.js"
+  ],
+  "scripts": {
+    "prepare": "node ../scripts/prepare.js"
+  },
+  "keywords": [
+    "storybook",
+    "aem",
+    "cli"
+  ],
+  "dependencies": {
+    "chalk": "^3.0.0",
+    "form-data": "^3.0.0",
+    "node-fetch": "^2.6.0",
+    "prompts": "^2.3.2",
+    "xml2json": "^0.12.0"
+  },
+  "devDependencies": {}
+}
