+++ cli/src/cmds/version.ts
@@ -1,6 +1,6 @@
 const { getInstalledVersion } = require('../utils/versionCheck');
 const log = require('../utils/logger');
 
-module.exports = async (args) => {
-    log(`Installed Version: ${getInstalledVersion()}`);
-}
+module.exports = async args => {
+  log(`Installed Version: ${getInstalledVersion()}`);
+};
