+++ cli/src/utils/createPageJCRContent.ts
@@ -1,0 +1,21 @@
+const fetchFromAEM = require('./fetchFromAEM');
+
+const createPageJCRContent = async config => {
+  const params = [
+    `jcr:primaryType=cq:PageContent`,
+    `sling:resourceType=${config.aemContentDefaultPageResourceType}`,
+    `title=${config.component}`,
+    `cq:template=${config.aemContentDefaultPageTemplate}`,
+  ];
+
+  const response = await fetchFromAEM({
+    url: `${config.aemContentPath}/${config.component}/jcr:content?${params.join('&')}`,
+    method: 'POST',
+    errorMessage: 'Error creating JCR:Content:',
+  });
+
+  if (await response.ok) return true;
+  return false;
+};
+
+module.exports = createPageJCRContent;
