--- cli/src/utils/error.js
@@ -1,4 +1,0 @@
-module.exports = (message, exit) => {
-    console.error(message)
-    exit && process.exit(1)
-}
