+++ cli/src/utils/error.ts
@@ -1,0 +1,4 @@
+export const error = (message, exit) => {
+  console.error(message);
+  exit && process.exit(1);
+};
