+++ cli/src/utils/getChoicesFromDirectories.ts
@@ -1,0 +1,20 @@
+const path = require('path');
+const getDirectories = require('./getDirectories');
+
+const cwd = process.cwd();
+
+module.exports = (config, directoryPath) => {
+  const directories = getDirectories(
+    path.resolve(cwd, config.projectRoot, config.relativeProjectRoot, directoryPath)
+  );
+  const choices = [];
+
+  directories.forEach(directory => {
+    choices.push({
+      title: directory,
+      value: directory,
+    });
+  });
+
+  return choices;
+};
