--- cli/src/utils/getDirectories.js
@@ -1,7 +1,0 @@
-const { readdirSync } = require('fs');
-
-module.exports = source => {
-    return readdirSync(source, { withFileTypes: true })
-                .filter(dirent => dirent.isDirectory())
-                .map(dirent => dirent.name);
-};
