+++ cli/src/utils/getDirectories.ts
@@ -1,0 +1,7 @@
+const { readdirSync } = require('fs');
+
+export const getDirectories = source => {
+  return readdirSync(source, { withFileTypes: true })
+    .filter(dirent => dirent.isDirectory())
+    .map(dirent => dirent.name);
+};
