--- cli/src/utils/logger.js
@@ -1,6 +1,0 @@
-const chalk = require('chalk');
-
-module.exports = (...args) => {
-    const sbaem = chalk.rgb(255, 71, 133).bold('[sb-aem]');
-    console.log(`${sbaem}`, ...args);
-};
