+++ scripts/compile-tsc.js
@@ -6,7 +6,11 @@ const shell = require('shelljs');
 function getCommand(watch) {
   const tsc = path.join(__dirname, '..', 'node_modules', '.bin', 'tsc');
 
-  const args = [`--outDir ${path.join(__dirname, '..', 'app/aem/dist')}`, '--listEmittedFiles true', '--allowjs true'];
+  const args = [
+    `--outDir ${path.join(__dirname, '..', 'app/aem/dist')}`,
+    '--listEmittedFiles true',
+    '--allowjs true',
+  ];
 
   if (watch) {
     args.push('-w');
